requirements.txt
comtypes==1.4.11
et_xmlfile==2.0.0
logging==0.4.9.6
MouseInfo==0.1.3
numpy==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
psutil==7.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywinauto==0.6.9
six==1.17.0
tzdata==2025.2

pytesseract~=0.3.13